module 1
This commit adds a new organization and structure for tidybiology.org. This adds a module.Rmd file while then calls several object.Rmd child chunks to string together the learning experience. This also adds helper fucntions to automatically create a table of objects and the modules they are included within. Because of this, the naming convention of objects is fixed.